requirements.txt
streamlit
pyttsx3
speechrecognition
wikipedia
groq
chromadb
python-dotenv







